Update scraperFreeWeibo.py
Updated the HotTerms variable. Before it was looking for the tag span in the content, however some posts may not have the span and it was being saved a null even though it is attached to a HotTerms. Now, when the Hot term is picked up during the initial scraping process of the Hot Term Table, that variable is passed to the next function and then send to the items.py page to be processed.